file conflict detection.

+++ PWSPlugin/src/edu/bpl/pwsplugin/acquisitionSequencer/factories/PauseFactory.java
@@ -116,8 +116,8 @@ class PauseStep extends EndpointStep {
     }
     
     @Override
-    public Integer numberNewAcqs() {
-        return 0;
+    public Double numberNewAcqs() {
+        return 0.0;
     }
 }
